requirements.txt
django
djangorestframework
psycopg2-binary
python-dotenv
djangorestframework-simplejwt
flake8
django-filter